requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
slowapi==0.1.8
motor==3.3.2
python-dotenv==1.0.0
asyncpg==0.29.0
python-jose>=3.3.0
passlib==1.7.4
